PKG fix & BUILD
Fix up PKG file extraction and file_handle handling
Add build script, update requirements, add pyinstaller

+++ ps/pkg/entry.py
@@ -5,6 +5,7 @@ from clint.textui import puts
 
 from ps.utils import DEFAULT_LOCAL_IO_BLOCK_SIZE, read_u32, read_u64, Endianess
 from .decryptor import DecryptorIO
+from .entry_type import EntryType
 
 
 class PkgEntry(object):
@@ -15,26 +16,36 @@ class PkgEntry(object):
         self.name_size: int = read_u32(f, endianess=Endianess.BIG_ENDIAN)
         self.file_offset: int = read_u64(f, endianess=Endianess.BIG_ENDIAN)
         self.file_size: int = read_u64(f, endianess=Endianess.BIG_ENDIAN)
-        self.type: int = read_u32(f, endianess=Endianess.BIG_ENDIAN)
+        entry_flags = read_u32(f, endianess=Endianess.BIG_ENDIAN)
+        # TODO: Use flags or do this better somehow
+        self.flag_overwrite: bool = (entry_flags >> 3 & 0x80) > 0
+        self.flag_psp: bool = (entry_flags >> 3 & 0x10) > 0
+        self.type: EntryType = EntryType(entry_flags & 0xFF)
         self.pad: int = read_u32(f, endianess=Endianess.BIG_ENDIAN)
-        puts("Name Offset: {}".format(self.name_offset))
-        puts("Name Size: {}".format(self.name_size))
-        puts("File Offset: {}".format(self.file_offset))
-        puts("File Size: {}".format(self.file_size))
-        puts("Type: {}".format(self.type))
-        puts("Pad: {}".format(self.pad))
+        puts(f"Name Offset: {self.name_offset}")
+        puts(f"Name Size: {self.name_size}")
+        puts(f"File Offset: {self.file_offset}")
+        puts(f"File Size: {self.file_size}")
+        puts(f"Type: {self.type}")
+        puts(f"Overwrite: {self.flag_overwrite}")
+        puts(f"PSP: {self.flag_psp}")
+        puts(f"Pad: {self.pad}")
 
         f.seek(self.name_offset, DecryptorIO.SEEK_DATA_OFFSET)
         # Not decoded to ASCII or UTF8 because some packages even though they are valid, have invalid characters
-        self.name: bytes = f.read(self.name_size)
-        puts("Name: {}".format(self.name))
+        self.name: str = f.read(self.name_size).decode('UTF-8')
+        puts(f"Name: {self.name}")
+
+    @property
+    def is_file(self):
+        return self.type.is_file
 
     @staticmethod
     def size():
         return 4 * 4 + 2 * 8
 
     def save_file(self, path: str, block_size: int = DEFAULT_LOCAL_IO_BLOCK_SIZE, use_package_path: bool = False,
-                  create_directories: bool = False) -> bool:
+                  create_directories: bool = True) -> bool:
         if (os.path.exists(path) and os.path.isdir(path)) or (not os.path.exists(path) and path.endswith(('/', '\\'))):
             if use_package_path:
                 path = os.path.join(path, self.name)
@@ -50,10 +61,17 @@ class PkgEntry(object):
         if create_directories and not os.path.exists(directory):
             os.makedirs(directory)
 
-        with open(path, 'wb') as export:
-            self.f.seek(self.file_offset, DecryptorIO.SEEK_DATA_OFFSET)
-            bytes_remaining: int = self.file_size
-            while bytes_remaining != 0:
-                to_read: int = block_size if bytes_remaining >= block_size else bytes_remaining
-                bytes_remaining -= export.write(self.f.read(to_read))
+        if self.is_file:
+            puts(f'Extracting file: {self.name} -> {path}')
+            if not os.path.exists(path) or self.flag_overwrite:
+                with open(path, 'wb') as export:
+                    self.f.seek(self.file_offset, DecryptorIO.SEEK_DATA_OFFSET)
+                    bytes_remaining: int = self.file_size
+                    while bytes_remaining != 0:
+                        to_read: int = block_size if bytes_remaining >= block_size else bytes_remaining
+                        bytes_remaining -= export.write(self.f.read(to_read))
+                    return True
+        else:
+            puts(f'Creating directory: {self.name} -> {path}')
+            os.makedirs(path)
             return True
